Preprocessor stuff removed
