+++ 1. Process/pipe2.hpp
@@ -1,0 +1,6 @@
+#ifndef PIPE2_H
+#define PIPE2_H
+
+int pipe2(int pipefd[2], int flags);
+
+#endif
