+++ 1. Process/process.cpp
@@ -1,5 +1,5 @@
 #include "process.hpp"
-#include "ostype.hpp"
+#include "pipe2.hpp"
 
 #include <cstdlib>
 #include <fcntl.h>
@@ -7,21 +7,6 @@
 #include <sys/wait.h>
 #include <unistd.h>
 #include <signal.h>
-
-#if PLATFORM_ENUM == OS_OSX
-    int pipe2(int pipefd[2], int flags) {
-        int ret_val = pipe(pipefd);
-        if (ret_val < 0)
-            return ret_val;
-        if (flags == O_CLOEXEC) {
-            if (fcntl(pipefd[0], F_SETFD, FD_CLOEXEC) == -1)
-                return -1;
-            if (fcntl(pipefd[1], F_SETFD, FD_CLOEXEC) == -1)
-                return -1;
-        }
-        return ret_val;
-    }
-#endif
 
 std::vector<char*> arg_list(const std::string &name, const std::vector <std::string>& args) {
     std::vector<char*> cstyle_args;
